fix(api,agora): fix email login blocked by Reacher and Enter key redirect
- Move Reacher email deliverability check inside authenticateEmailAttempt
  so it only runs for registration, not login/merge (Reacher was blocking
  all email auth when unavailable)
- Reduce Reacher timeout from 30s to 5s to fit within frontend's 8s window
- Add type="button" to ZKGradientButton to prevent Enter key from
  triggering the first submit button (was redirecting to zkpassport)
- Add SMTP timeout and request logging config for Reacher in production

Deploy: api, agora
