+++ script/docker-compose-production.yml
@@ -60,6 +60,8 @@ services:
       RCH__HTTP_HOST: "0.0.0.0"
       RCH__FROM_EMAIL: "noreply@probe.agoracitizen.network"
       RCH__HELLO_NAME: "probe.agoracitizen.network"
+      RCH__SMTP_TIMEOUT: "5"
+      RUST_LOG: "reacher=info"
     expose:
       - "8080"
     restart: unless-stopped
